refactor: Restructure header components and update imports for improved organization
