--- src/features/website/header/index.ts
@@ -1,1 +1,0 @@
-export { Header } from './view/Header';
